add read from csv if else

+++ acquire.py
@@ -24,25 +24,29 @@ def get_df(name):
     returns a df containing all pages and
     creates a .csv file for future use.
     """
-    base_url = 'https://python.zach.lol'
-    api_url = base_url + '/api/v1/'
-    response = requests.get(api_url + name)
-    data = response.json()
-    
-    # create list from 1st page
-    my_list = data['payload'][name]
-    
-    # loop through the pages and add to list
-    while data['payload']['next_page'] != None:
-        response = requests.get(base_url + data['payload']['next_page'])
+
+    if os.path.isfile(name + '.csv'):
+        df = pd.read_csv((name + '.csv'), index_col=0)
+    else:
+        base_url = 'https://python.zach.lol'
+        api_url = base_url + '/api/v1/'
+        response = requests.get(api_url + name)
         data = response.json()
-        my_list.extend(data['payload'][name])
-    
-    # Create DataFrame from list
-    df = pd.DataFrame(my_list)
-    
-    # Write DataFrame to csv file for future use
-    df.to_csv(name + '.csv')
+        
+        # create list from 1st page
+        my_list = data['payload'][name]
+        
+        # loop through the pages and add to list
+        while data['payload']['next_page'] != None:
+            response = requests.get(base_url + data['payload']['next_page'])
+            data = response.json()
+            my_list.extend(data['payload'][name])
+        
+        # Create DataFrame from list
+        df = pd.DataFrame(my_list)
+        
+        # Write DataFrame to csv file for future use
+        df.to_csv(name + '.csv')
     return df
 
 def merge_items_stores_sales(sales_df, stores_df, items_df):
